requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
celery>=5.3.0
redis>=4.6.0
python-multipart>=0.0.6
jinja2>=3.1.2
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
edge-tts>=6.1.9
webvtt-py>=0.4.6
pydub>=0.25.1
mutagen>=1.47.0
Pillow>=10.0.0
num2words>=0.5.13
rich>=13.0.0
EbookLib>=0.18
html2text
pedalboard>=0.8.0
google-api-python-client>=2.93.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
